Relocated most of the replayio async code to shared/async
